pyproject.toml
[tool.poetry]
name = "opset"
version = "4.0.0"
description = "A library for simplifying the configuration of Python applications at all stages of deployment."
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/MarcDufresne/opset"
repository = "https://github.com/MarcDufresne/opset"
documentation = "https://github.com/MarcDufresne/opset"
keywords = ["config", "management", "configuration", "logging", "setup"]
authors = ["Marc-André Dufresne <marc.andre.dufresne@gmail.com>", "Alexandre Jutras <jutras.alexandre@gmail.com >"]
packages = [
    {include = "opset"}
]

[tool.poetry.dependencies]
python = "^3.10"
munch = "^4.0"
pytz = ">=2023.3"
pyyaml = "^6.0"
structlog = "^23.1"
colorama = "^0.4"
google-cloud-secret-manager = {version = "^2.16.4",  optional = true}
setuptools = "^68.2.2"
pydantic = "^2.4.2"

[tool.poetry.extras]
gcp = ["google-cloud-secret-manager"]

[tool.poetry.group.dev.dependencies]
coveralls = "^3.3.1"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-sugar = "^0.9.7"
black = "^23.7.0"
mypy = "^1.4.1"
ruff = "^0.0.278"
types-setuptools = "^68.0.0.2"
types-pyyaml = "^6.0.12.10"
pytest-mock = "^3.11.1"

[tool.ruff]
line-length = 120

select = [
    "F",
    "E",
    "T20",
    "I",
    "T10"
]
ignore = ["E501"]
unfixable = ["F841"]

[tool.ruff.per-file-ignores]
"tests/*" = ["N802", "N803", "E203"]

[tool.black]
line-length = 120

[tool.mypy]
python_version = 3.11
namespace_packages = true
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
